GranjaWindowsApp/requirements.txt
PyQt6>=6.0.0
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
